+++ Bike/logic/sensor.h
@@ -43,7 +43,7 @@
 // *************************************************************************************************
 // Prototypes section
 extern void reset_sensor(void);
-extern u32 sensor_get_distance(void);
+extern u16 sensor_get_distance(void);
 
 // *************************************************************************************************
 // Defines section
